Create BMI Calculator
 Program in Python to compute the BMI of a person and display the risk associated with it by entering the height in 'm' and weight in' kg'. Refer to the following table and code accordingly.

Note:

To calculate the BMI apply the formula:  BMI = weight(kg)/( height(m)*height(m) ).    
The result must be adjusted to one decimal place. 
When the height or weight is entered as a negative number or zero, then display the message "Provide a valid input" and stop the program.


BMI-Risk

27.5 and above-High Risk

23 - 27.4-Moderate Risk

18.5 - 22.9-Low Risk

Below 18.5-Risk of nutritional deficiency diseases

Sample input statement 1:
Enter the weight of the person(kg): 85
Enter the height of the person(m): 1.75

Sample output statement 1:
Your BMI is 27.8 (High Risk)

Sample input statement 2:
Enter the weight of the person(kg): 0
Enter the height of the person(m): 1.58

Sample output statement 2:
Provide a valid input


Sample input statement 3:
Enter the weight of the person(kg): 80
Enter the height of the person(m): -1

Sample output statement 3:
Provide a valid input